Use ConstructorArgs trait for args on deploy with constructor
